Move matching methods from tasks.py to new match.py
