+++ seed/data_importer/tests/integration/test_equivalence_partioner_with_file.py
@@ -10,7 +10,7 @@ import os.path as osp
 
 from django.core.files.uploadedfile import SimpleUploadedFile
 
-from seed.data_importer import tasks
+from seed.data_importer import tasks, match
 from seed.data_importer.tests.util import (
     FAKE_MAPPINGS,
 )
@@ -47,7 +47,7 @@ class TestEquivalenceWithFile(DataMappingBaseTestCase):
 
     def test_equivalence(self):
         all_unmatched_properties = self.import_file.find_unmatched_property_states()
-        unmatched_properties, duplicate_property_states = tasks.filter_duplicated_states(
+        unmatched_properties, duplicate_property_states = match.filter_duplicated_states(
             all_unmatched_properties
         )
         partitioner = EquivalencePartitioner.make_propertystate_equivalence()
